+++ ltr/data/data_set_unittest.cc
@@ -46,7 +46,7 @@ TEST_F(DataSetTest, DataSetLightSubsetTest) {
   data_set.add(o3);
 
   ltr::DataSet<ltr::Object> light_subset = ltr::utility::lightSubset(data_set,
-      std::vector<size_t>(1, 1));
+      std::vector<int>(1, 1));
 
   EXPECT_EQ(light_subset.at(0), o2);
 }
